refactor: split public event service responsibilities
